2010-01-12  Neal H. Walfield  <neal@gnu.org>
	* src/lyrics.c: Fix warnings.
	* src/musicdb.h: Likewise.
	* src/musicdb.c: Likewise.
	* src/playlist.c: Likewise.
	* src/playlist.h: Likewise.
	* src/playlists.c: Likewise.
	* src/search.y: Likewise.
	* src/starling.c: Likewise.


git-svn-id: svn://projects.linuxtogo.org/svn/gpe/trunk@10046 d4735bf7-8e1f-0410-bce4-ef584eb11c25

+++ extra/starling/src/musicdb.c
@@ -1496,7 +1496,7 @@ music_db_set_info_from_tags (MusicDB *db, int uid, GstTagList *tags)
 
 struct info_callback_data
 {
-  int (*user_callback) (int uid, struct music_db_info *info);
+  int (*user_callback) (guint uid, struct music_db_info *info);
   int ret;
   /* The number of leading columns to ignore.  */
   int ignore;
@@ -1516,7 +1516,7 @@ info_callback (void *arg, int argc, char **argv, char **names)
   if (argv[0])
     info.aggregate = atoi (argv[0]);
 
-  int uid = atoi (argv[1]);
+  guint uid = atoll (argv[1]);
 
   info.fields = 0;
 
@@ -1554,7 +1554,7 @@ info_callback (void *arg, int argc, char **argv, char **names)
 
 int
 music_db_for_each (MusicDB *db, const char *list,
-		   int (*user_callback) (int uid, struct music_db_info *info),
+		   int (*user_callback) (guint uid, struct music_db_info *info),
 		   enum mdb_fields *order,
 		   enum mdb_fields scope,
 		   const char *constraint)
